+++ django/API/views/server_info.py
@@ -1,11 +1,15 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 # Copyright (C) 2022 Edith Coates
+# Copyright (C) 2022 Colin B. Macdonald
 
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.exceptions import APIException
 from rest_framework import status
 from django.core.exceptions import ObjectDoesNotExist
+
+from plom import __version__
+from plom import Plom_API_Version
 
 from Mark.services import MarkingTaskService
 from Papers.services import SpecificationService
@@ -41,7 +45,7 @@ class ServerVersion(APIView):
     """
 
     def get(self, request):
-        version = "Plom server version 0.12.0.dev with API 55"
+        version = f"Plom server version {__version__} with API {Plom_API_Version}"
         return Response(version)