+++ plom/db/examDB.py
@@ -192,8 +192,7 @@ class PlomDB:
         add_or_change_id_prediction,
         remove_id_prediction,
         remove_id_from_paper,
-        hasNoAnswerRubric,
-        createNoAnswerRubric,
+        hasAutoGenRubrics,
     )
 
     from plom.db.db_upload import (